+++ src/imio/project/pst/browser/viewlets.py
@@ -126,15 +126,7 @@ class ContextInformationViewlet(MessagesViewlet):
             )
 
         if self.context.portal_type == "operationalobjective":
-            act_planned_end_date = [
-                act.planned_end_date
-                for act in api.content.find(
-                    context=self.context,
-                    portal_type=["pstaction", "action_link", "pstsubaction", "subaction_link"],
-                )
-                if act.planned_end_date
-            ]
-            if act_planned_end_date:
+            if self.context.get_actions_planned_end_dates():
                 if not self.context.planned_end_date:
                     msg = _(u"The planned end date field of this operational objective has no value, "
                             u"the furthest planned end date date from its actions has been taken")
@@ -146,7 +138,7 @@ class ContextInformationViewlet(MessagesViewlet):
                             can_hide=False,
                         )
                     )
-                elif max(act_planned_end_date) > self.context.planned_end_date:
+                elif self.context.get_max_actions_planned_end_dates() > self.context.planned_end_date:
                     msg = _(u"The planned end date of any one of the actions is greater than the planned end date "
                             u"of the operational objective")
                     ret.append(